get rids of the DefaultError
